+++ src/MinionStrategy/OpponentExpression.java
@@ -3,7 +3,7 @@ import Game.Minion;
 
 public class OpponentExpression extends InfoExpression{
     @Override
-    public int evaluate(Minion minion) throws Exception {
+    public long evaluate(Minion minion) throws Exception {
         return minion.getOpponent();
     }
